kafka-consumer/requirements.txt
pymongo==4.3.3
pandas==2.2.3
python-dotenv==1.0.0
pytz==2023.3
confluent-kafka>=2.2.0
